requirements.txt
psutil
mysql-connector-python
python-dotenv
tabulate
numpy
pandas